hotfix in .h changed uint8_t to int

+++ examples/BasicTempOut/LTS01A_MAX31725.h
@@ -12,7 +12,7 @@
 class MAX31725
 {
 private:     
-    uint8_t _i2cAddress;
+    int _i2cAddress;
 
 public:
     // Konstruktor
